Convert Dashboard to Rails Engine
