+++ app/assets/javascripts/bandit/application.js
@@ -1,0 +1,3 @@
+//= require ./jquery.min
+//= require ./highstock
+//= require ./bandit
