+++ app/assets/javascripts/bandit/bandit.js
