+++ app/assets/javascripts/bandit/highstock.js
