+++ app/assets/stylesheets/bandit/application.css
@@ -1,0 +1,5 @@
+/*
+ *= require ./toupee/reset
+ *= require ./base
+ *= require_tree ./toupee
+*/
