+++ app/assets/stylesheets/bandit/toupee/buttons.css
