+++ app/assets/stylesheets/bandit/toupee/structure.css
