+++ app/controllers/bandit/application_controller.rb
@@ -1,0 +1,4 @@
+module Bandit
+  class ApplicationController < ActionController::Base
+  end
+end
