+++ config/routes.rb
@@ -1,0 +1,5 @@
+Bandit::Engine.routes.draw do
+  get ':id' => 'dashboard#show', :as => 'bandit'
+
+  root :to => 'dashboard#index'
+end
