+++ lib/bandit/engine.rb
@@ -1,0 +1,6 @@
+module Bandit
+  class Engine < Rails::Engine
+    isolate_namespace Bandit
+    engine_name 'bandit'
+  end
+end
