--- lib/generators/bandit/dashboard_generator.rb
@@ -1,31 +1,0 @@
-module Bandit
-  module Generators
-
-    class DashboardGenerator < Rails::Generators::Base
-      desc "Create a bandit dashboard controller"
-      source_root File.expand_path('../templates', __FILE__)
-
-      def copy_controller
-        copy_file 'bandit_controller.rb', 'app/controllers/bandit_controller.rb'
-      end
-      
-      def copy_view
-        directory 'dashboard/view', 'app/views/bandit'
-        copy_file 'dashboard/bandit.html.erb', 'app/views/layouts/bandit.html.erb'
-        directory 'dashboard/helpers', 'app/helpers'
-      end
-
-      def copy_assets
-        directory 'dashboard/js', 'vendor/assets/javascripts/bandit'
-        directory 'dashboard/css', 'vendor/assets/stylesheets/bandit'
-      end
-
-      def message
-        say "\n\tNow, add the following to your config/routes.rb file:"
-        say "\t\tresources :bandit\n\n"
-      end
-
-   end
-
-  end
-end
