--- lib/generators/bandit/templates/dashboard/bandit.html.erb
@@ -1,43 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" lang="en">
-  <head>
-  <title>Bandit Dashboard</title>
-  <%= stylesheet_link_tag "bandit/application.css" %>
-  <%= javascript_include_tag "bandit/jquery.min.js" %>
-  <%= javascript_include_tag "bandit/highstock.js" %>
-  <%= javascript_include_tag "bandit/bandit.js" %>
-  </head>
-  <body>
-    <header>
-      <div class="container">
-	<div class="span-10 brand">
-	  <h1><%= link_to "Bandit Dashboard", :controller => 'bandit', :action => 'index' %></h1>
-	</div>
-	<nav class="span-14 last">
-	  <ul>
-            <li><%= link_to "index", :controller => 'bandit', :action => 'index' %></li>
-	  </ul>
-	</nav>
-      </div>
-    </header>
-    <div class="container">
-      <%= yield %>
-      <aside class='span-3 last'>
-	<section class='mod'>
-	  <header>
-	    <h2>Experiments</h2>
-	  </header>
-	  <article>
-	    <ul>
-	      <% Bandit.experiments.each { |experiment| %>
-	      <li><%= link_to experiment.name.to_s.titleize, url_for(:action => 'show', :id => experiment.name) %></li>
-	      <% } %>
-	    </ul>
-	  </article>
-	</section>
-      </aside>
-    </div>
-    <footer>
-    </footer>
-  </body>
-</html>
