--- lib/generators/bandit/templates/dashboard/css/application.css
@@ -1,7 +1,0 @@
-@import url(toupee/reset.css);
-@import url(base.css);
-@import url(toupee/buttons.css);
-@import url(toupee/forms.css);
-@import url(toupee/modules.css);
-@import url(toupee/structure.css);
-@import url(toupee/typography.css);
