--- lib/generators/bandit/templates/dashboard/view/show.html.erb
@@ -1,19 +1,0 @@
-<section class='span-20 colborder'>
-  <div class='title'><h1><%= @experiment.name.to_s.titleize %></h1></div>
-  <p>
-    <%= @experiment.description %>
-  </p>
-
-  <%= render :partial => 'experiment_table', :locals => { :experiment => @experiment } %>
-
-  <div id="totals_gcontainer" class="graph"></div>
-
-  <div id="percents_gcontainer" class="graph"></div>
-
-</section>
-
-<script type="text/javascript">
-var url = "<%= url_for(:controller => 'bandit', :action => 'show', :id => @experiment.name, :format => 'csv') %>";
-var title = "<%= @experiment.name.to_s.titleize %>";
-show_chart(title, url);
-</script>
